Assistant checkpoint: Update user titles and types in constants file
Assistant generated file changes:
- client/src/lib/constants.ts: Update user titles and information

---

User prompt:

All sources of user information do not seem synchronized. Anthony Zachria's title is 'DO' which is correct in the Administration > Users and is not correct in the list you provided. Todd Want's title is 'APP' not MD. Does the administration user information sync to other destinations of user information?